+++ tests/Checks/MeilisearchCheckTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Http;
-use Spatie\Health\Checks\Checks\MeiliSearchCheck;
+use Spatie\Health\Checks\Checks\MeilisearchCheck;
 use Spatie\Health\Enums\Status;
 
 it('will determine that a working meilisearch is ok', function () {
@@ -9,7 +9,7 @@ it('will determine that a working meilisearch is ok', function () {
         '*' => Http::response(['status' => 'available']),
     ]);
 
-    $result = MeiliSearchCheck::new()->run();
+    $result = MeilisearchCheck::new()->run();
 
     expect($result->status)->toBe(Status::ok());
 });
@@ -19,7 +19,7 @@ it('will determine that another status is not ok', function () {
         '*' => Http::response(['status' => 'not ok']),
     ]);
 
-    $result = MeiliSearchCheck::new()->run();
+    $result = MeilisearchCheck::new()->run();
 
     expect($result->status)->toBe(Status::failed());
 });
@@ -29,7 +29,7 @@ it('will determine that an http error is not ok', function () {
         '*' => Http::response(status: 500),
     ]);
 
-    $result = MeiliSearchCheck::new()->run();
+    $result = MeilisearchCheck::new()->run();
 
     expect($result->status)->toBe(Status::failed());
 });